MDL-67709 user: Fix for Custom Profile Fields index.php
We discovered that the User Profile Fields index.php (user/profile/index.php) page loads extremely slow on our Moodle LMS when we increased the number of custom profile fields on the LMS.

Our LMS has about 80 custom profile fields with about 25,000 users. The page progressively got slower the more fields and users we added onto the site until it started taking over 80+ seconds to load the index page.

After some further investigation into the index.php code we discovered a line of code that is not used that add's a huge load to the page because it makes a DB call in a nested foreach loop on each custom profile field.

The fix is to remove this unused variable.